+++ server/src/test/java/org/apache/seata/server/coordinator/DefaultCoordinatorTest.java
@@ -27,6 +27,7 @@ import io.netty.channel.Channel;
 import org.apache.seata.common.DefaultValues;
 import org.apache.seata.common.XID;
 import org.apache.seata.common.loader.EnhancedServiceLoader;
+import org.apache.seata.common.store.SessionMode;
 import org.apache.seata.common.util.NetUtil;
 import org.apache.seata.common.util.ReflectionUtil;
 import org.apache.seata.config.Configuration;
@@ -45,7 +46,6 @@ import org.apache.seata.core.rpc.processor.RemotingProcessor;
 import org.apache.seata.server.metrics.MetricsManager;
 import org.apache.seata.server.session.GlobalSession;
 import org.apache.seata.server.session.SessionHolder;
-import org.apache.seata.server.store.StoreConfig.SessionMode;
 import org.apache.seata.server.util.StoreUtil;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
