+++ server/src/test/java/org/apache/seata/server/raft/execute/GlobalSessionExecuteTest.java
@@ -16,6 +16,8 @@
  */
 package org.apache.seata.server.raft.execute;
 
+import org.apache.seata.common.store.LockMode;
+import org.apache.seata.common.store.SessionMode;
 import org.apache.seata.common.util.NetUtil;
 import org.apache.seata.config.ConfigurationCache;
 import org.apache.seata.core.exception.TransactionException;
@@ -30,7 +32,6 @@ import org.apache.seata.server.session.GlobalSession;
 import org.apache.seata.server.session.SessionHolder;
 import org.apache.seata.server.session.SessionManager;
 import org.apache.seata.server.storage.SessionConverter;
-import org.apache.seata.server.store.StoreConfig;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
@@ -47,9 +48,9 @@ class GlobalSessionExecuteTest {
     @BeforeAll
     public static void setUp(ApplicationContext context){
         System.setProperty("server.raft.serverAddr", NetUtil.getLocalIp() + ":9091");
-        SessionHolder.init(StoreConfig.SessionMode.RAFT);
+        SessionHolder.init(SessionMode.RAFT);
         LockerManagerFactory.destroy();
-        LockerManagerFactory.init(StoreConfig.LockMode.RAFT);
+        LockerManagerFactory.init(LockMode.RAFT);
     }
 
     @AfterAll
