+++ server/src/test/java/org/apache/seata/server/session/SessionHolderTest.java
@@ -19,9 +19,8 @@ package org.apache.seata.server.session;
 import java.io.File;
 import java.io.IOException;
 import org.apache.seata.common.XID;
+import org.apache.seata.common.store.SessionMode;
 import org.apache.seata.core.constants.ConfigurationKeys;
-import org.apache.seata.server.store.StoreConfig.SessionMode;
-import org.apache.seata.server.session.SessionHolder;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
